Port tc39 tests
